+++ pkg/product/routes/find_one.go
@@ -13,7 +13,7 @@ func FineOne(ctx *gin.Context, c pb.ProductServiceClient) {
 	id, _ := strconv.ParseInt(ctx.Param("id"), 10, 32)
 
 	res, err := c.FindOne(context.Background(), &pb.FindOneRequest{
-		Id: int32(id),
+		Id: int64(id),
 	})
 
 	if err != nil {
